db: Major reorganization of the query code
I developed a plan by which queries are constructed in a uniform fashion,
then reworked all the query code to use it.  This removes a lot of duplicate
strings and boiler plate and makes it somewhat easier to manage.  Also,
straightening out the queries speeds up FM quite a bit.