db: Write a fast method to compute nr_extents and travel_score.
Now we can make the fiemap engine invoke it by default so we can
assume the values will always be there.  This enables us to reduce
runtime lag because we no longer try to compute it query time.
Plumb the resulting mess into a CLI command; fix a CLI bug; and
make it easier to trace SQL.